+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+### Changed
+- **Breaking** The `Picker::restart` method clears the query string.
+- Migrated to Rust Edition 2024.
+
+### Fixed
+- Calling `Picker::update_query` correctly propagates the update to the internal matching engine.
+
 ## [0.8.1] - 2025-02-07
 
 ### Added